% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tutorial_data.R
\name{setup_tutorial_data}
\alias{setup_tutorial_data}
\title{Setup tutorial data by downloading from GitHub}
\usage{
setup_tutorial_data(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If TRUE, provides detailed progress information. Default is FALSE.}
}
\value{
Invisible NULL. Called for side effects.
}
\description{
Downloads data directories from GitHub repository for installed tutorials
}
